prjctShop/requirements.txt
amqp==5.3.1
asgiref==3.9.2
billiard==4.2.2
celery==5.4.0
certifi==2025.10.5
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cron_descriptor==2.0.6
Django==5.1
django-allauth==64.1.0
django-celery-beat==2.8.1
django-cors-headers==4.3.1
django-htmx==1.22.0
django-taggit==6.1.0
django-timezone-field==7.1
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.11
kombu==5.5.4
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stripe==13.0.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.7.0
